Remove dead file upload actions, reorganize modules.
