+++ client/src/action/uploadBallotManifest.ts
@@ -1,32 +1,90 @@
+import action from '.';
+
 import { endpoint } from '../config';
 
-import createFileUploadAction from './createFileUploadAction';
+const importUrl = endpoint('import-ballot-manifest');
+const uploadUrl = endpoint('upload-file');
 
 
-const url = endpoint('upload-ballot-manifest');
-
-
-function createFormData(countyId: number, file: Blob, hash: string) {
+function createFormData(file: Blob, hash: string) {
     const formData = new FormData();
 
-    formData.append('county', `${countyId}`);
-    formData.append('bmi_file', file);
+    formData.append('file', file);
     formData.append('hash', hash);
 
     return formData;
 }
 
-function createSent(countyId: number, file: Blob, hash: string) {
-    return { countyId, file, hash };
+async function importBallotManifest(body: any) {
+    const init: any = {
+        body: JSON.stringify(body),
+        credentials: 'include',
+        method: 'post',
+    };
+
+    try {
+        action('NEXT_IMPORT_BALLOT_MANIFEST_SEND');
+
+        const r = await fetch(importUrl, init);
+
+        const received = await r.json();
+        const sent = body;
+        const data = { received, sent };
+
+        if (!r.ok) {
+            action('NEXT_IMPORT_BALLOT_MANIFEST_FAIL', data);
+            return;
+        }
+
+        action('NEXT_IMPORT_BALLOT_MANIFEST_OK', data);
+    } catch (e) {
+        if (e.message === 'Failed to fetch') {
+            action('NEXT_IMPORT_BALLOT_MANIFEST_NETWORK_FAIL');
+        }
+
+        action('INTERNAL_ERROR');
+
+        throw e;
+    }
+}
+
+async function uploadBallotManifest(countyId: number, file: Blob, hash: string) {
+    const formData = createFormData(file, hash);
+
+    const init: any = {
+        body: formData,
+        credentials: 'include',
+        method: 'post',
+    };
+
+    try {
+        action('NEXT_UPLOAD_BALLOT_MANIFEST_SEND');
+
+        const r = await fetch(uploadUrl, init);
+
+        const received = await r.json();
+        const sent = { file, hash };
+        const data = { received, sent };
+
+        if (!r.ok) {
+            action('NEXT_UPLOAD_BALLOT_MANIFEST_FAIL', data);
+            return;
+        }
+
+        action('NEXT_UPLOAD_BALLOT_MANIFEST_OK', data);
+
+        importBallotManifest(received);
+    } catch (e) {
+        if (e.message === 'Failed to fetch') {
+            action('NEXT_UPLOAD_BALLOT_MANIFEST_NETWORK_FAIL');
+        }
+
+        action('INTERNAL_ERROR');
+
+        throw e;
+    }
+
 }
 
 
-export default createFileUploadAction({
-    createFormData,
-    createSent,
-    failType: 'UPLOAD_BALLOT_MANIFEST_FAIL',
-    networkFailType: 'UPLOAD_BALLOT_MANIFEST_NETWORK_FAIL',
-    okType: 'UPLOAD_BALLOT_MANIFEST_OK',
-    sendType: 'UPLOAD_BALLOT_MANIFEST_SEND',
-    url,
-});
+export default uploadBallotManifest;
