+++ client/src/action/uploadCvrExport.ts
@@ -1,32 +1,90 @@
+import action from '.';
+
 import { endpoint } from '../config';
 
-import createFileUploadAction from './createFileUploadAction';
+const importUrl = endpoint('import-cvr-export');
+const uploadUrl = endpoint('upload-file');
 
 
-const url = endpoint('upload-cvr-export');
-
-
-function createFormData(countyId: number, file: Blob, hash: string) {
+function createFormData(file: Blob, hash: string) {
     const formData = new FormData();
 
-    formData.append('county', `${countyId}`);
-    formData.append('cvr_file', file);
+    formData.append('file', file);
     formData.append('hash', hash);
 
     return formData;
 }
 
-function createSent(countyId: number, file: Blob, hash: string) {
-    return { countyId, file, hash };
+async function importCvrExport(body: any) {
+    const init: any = {
+        body: JSON.stringify(body),
+        credentials: 'include',
+        method: 'post',
+    };
+
+    try {
+        action('NEXT_IMPORT_CVR_EXPORT_SEND');
+
+        const r = await fetch(importUrl, init);
+
+        const received = await r.json();
+        const sent = body;
+        const data = { received, sent };
+
+        if (!r.ok) {
+            action('NEXT_IMPORT_CVR_EXPORT_FAIL', data);
+            return;
+        }
+
+        action('NEXT_IMPORT_CVR_EXPORT_OK', data);
+    } catch (e) {
+        if (e.message === 'Failed to fetch') {
+            action('NEXT_IMPORT_CVR_EXPORT_NETWORK_FAIL');
+        }
+
+        action('INTERNAL_ERROR');
+
+        throw e;
+    }
+}
+
+async function uploadCvrExport(countyId: number, file: Blob, hash: string) {
+    const formData = createFormData(file, hash);
+
+    const init: any = {
+        body: formData,
+        credentials: 'include',
+        method: 'post',
+    };
+
+    try {
+        action('NEXT_UPLOAD_CVR_EXPORT_SEND');
+
+        const r = await fetch(uploadUrl, init);
+
+        const received = await r.json();
+        const sent = { file, hash };
+        const data = { received, sent };
+
+        if (!r.ok) {
+            action('NEXT_UPLOAD_CVR_EXPORT_FAIL', data);
+            return;
+        }
+
+        action('NEXT_UPLOAD_CVR_EXPORT_OK', data);
+
+        importCvrExport(received);
+    } catch (e) {
+        if (e.message === 'Failed to fetch') {
+            action('NEXT_UPLOAD_CVR_EXPORT_NETWORK_FAIL');
+        }
+
+        action('INTERNAL_ERROR');
+
+        throw e;
+    }
+
 }
 
 
-export default createFileUploadAction({
-    createFormData,
-    createSent,
-    failType: 'UPLOAD_CVR_EXPORT_FAIL',
-    networkFailType: 'UPLOAD_CVR_EXPORT_NETWORK_FAIL',
-    okType: 'UPLOAD_CVR_EXPORT_OK',
-    sendType: 'UPLOAD_CVR_EXPORT_SEND',
-    url,
-});
+export default uploadCvrExport;
